add: renderMain.js
